生成器的后台-后端代码重构

+++ vince-generator-web-backend/src/main/java/com/vince/web/common/PageRequest.java
@@ -6,7 +6,8 @@ import lombok.Data;
 /**
  * 分页请求
  *
- * @author vince
+ * @author <a href="https://github.com/liyupi">程序员鱼皮</a>
+ * @from <a href="https://yupi.icu">编程导航知识星球</a>
  */
 @Data
 public class PageRequest {
@@ -14,12 +15,12 @@ public class PageRequest {
     /**
      * 当前页号
      */
-    private long current = 1;
+    private int current = 1;
 
     /**
      * 页面大小
      */
-    private long pageSize = 10;
+    private int pageSize = 10;
 
     /**
      * 排序字段
